Fix format specifier for 64-bit index in error messages
Commit e2ab21f2 ("Implement GL_EXT_shader_64bit_indexing") made the
index int64_t but left the "index out of range" messages using '%d'.

32-bit ARM and PowerPC align 64-bit variadic arguments, so '%d' reads
the skipped padding (a stack slot on ARM, a register on PowerPC) and
prints garbage. This breaks the runtests single-thread vs. multithread
output comparison.

Fixes #4180

+++ glslang/MachineIndependent/ParseContextBase.cpp
@@ -290,32 +290,32 @@ void TParseContextBase::checkIndex(const TSourceLoc& loc, const TType& type, int
                type.getArraySizes()->getOuterNode()->getAsSymbolNode() == nullptr; };
 
     if (index < 0) {
-        error(loc, "", "[", "index out of range '%d'", index);
+        error(loc, "", "[", "index out of range '%lld'", (long long)index);
         index = 0;
     } else if (type.isArray()) {
         if (type.isSizedArray() && !sizeIsSpecializationExpression() &&
             index >= type.getOuterArraySize()) {
-            error(loc, "", "[", "array index out of range '%d'", index);
+            error(loc, "", "[", "array index out of range '%lld'", (long long)index);
             index = type.getOuterArraySize() - 1;
         }
     } else if (type.isVector()) {
         if (index >= type.getVectorSize()) {
-            error(loc, "", "[", "vector index out of range '%d'", index);
+            error(loc, "", "[", "vector index out of range '%lld'", (long long)index);
             index = type.getVectorSize() - 1;
         }
     } else if (type.isMatrix()) {
         if (index >= type.getMatrixCols()) {
-            error(loc, "", "[", "matrix index out of range '%d'", index);
+            error(loc, "", "[", "matrix index out of range '%lld'", (long long)index);
             index = type.getMatrixCols() - 1;
         }
     } else if (type.isCoopVecNV()) {
         if (index >= type.computeNumComponents()) {
-            error(loc, "", "[", "cooperative vector index out of range '%d'", index);
+            error(loc, "", "[", "cooperative vector index out of range '%lld'", (long long)index);
             index = type.computeNumComponents() - 1;
         }
     } else if (type.isLongVector()) {
         if (!type.hasSpecConstantVectorComponents() && index >= type.computeNumComponents()) {
-            error(loc, "", "[", "vector index out of range '%d'", index);
+            error(loc, "", "[", "vector index out of range '%lld'", (long long)index);
             index = type.computeNumComponents() - 1;
         }
     }
